Experiments implementing sorting algorithms.
While learning about templates, thought it might be fun to try and implement a couple of sorting algorithms to help understand the use of the algorithm. Only had time to do insertion sort and quick sort. C++ is actually quite fun, but I'd still say I'm more of a C person.